requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai==0.3.2
PyMuPDF==1.23.8
aiohttp==3.9.1
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
beautifulsoup4==4.12.2
